Allow timeouts to be overridden on a per-request basis
Note that the timeout for network reads and writes is derived from the
configured timeout, not the request timeout. The request timeout only
modifies the parameter sent to the daemon.

This update needs to be simultaneous with the core change, to avoid a
fatal error due to the prototype change.

Bug: T250248
Depends-On: I1f410cd5d83588e584b6d27d2e106465f0fad23e
Change-Id: I350225b1f1f74b7eb196086927731d1955145d91