RELEASE doesn't take an argument
The PoolCounter server will accept an argument to the RELEASE command,
in that it will happily read it off the wire and then do nothing with it
at all.  RELEASE always releases whichever lock the client most recently
acquired (or none at all, if that client holds no locks).

Having the client code reflect this seems like the least confusing option.

Change-Id: I933a37d14c6086ae66d8e456ad01fcbf7c4f0a1d